move prioritization tests into separate module
